Moving CreatedCharacter state in App.js
